@@ -44,19 +44,18 @@
   },
   "homepage": "https://javier.xyz/morphin/",
   "devDependencies": {
-    "eslint": "^8.4.1",
-    "eslint-plugin-react": "^7.27.1",
+    "eslint": "^8.6.0",
+    "eslint-plugin-react": "^7.28.0",
     "gh-pages": "^3.2.3"
   },
   "dependencies": {
-    "base64-image-utils": "^1.2.0",
+    "canvas-image-utils": "^2.1.1",
     "capsize": "^2.0.0",
-    "jbx": "^1.6.0",
+    "jbx": "^1.7.5",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0-beta-24dd07bd2-20211208",
+    "react-dom": "^18.0.0-beta-24dd07bd2-20211208",
     "react-scripts": "^5.0.0",
-    "styled-components": "^5.3.3",
-    "tinycolor2": "^1.4.2"
+    "styled-components": "^5.3.3"
   }
 }
